requirements.txt
# Star-Daemon Requirements
# Updated: 2025-10-26
# 
# For Shai-Hulud mitigations with hashes, run:
# pip-compile --generate-hashes requirements.in -o requirements-lock.txt
# pip install -r requirements-lock.txt --require-hashes

# Core dependencies
python-dotenv==1.1.1
requests==2.32.5

# GitHub API
PyGithub==2.8.1

# Mastodon
Mastodon.py==2.1.4

# BlueSky / AT Protocol
atproto==0.0.62

# Discord
discord.py==2.6.4
discord-webhook==1.3.1

# Matrix
matrix-nio==0.25.2

# Security and utilities
cryptography==46.0.3
certifi==2025.10.5
urllib3==2.5.0
charset-normalizer==3.4.4
idna==3.11
protobuf==4.25.8

# Async support for Matrix
aiohttp==3.13.1
aiofiles==25.1.0

# Optional: Doppler SDK for secrets management
doppler-sdk==1.3.0

# Dependencies
python-dateutil==2.9.0.post0
six==1.17.0
beautifulsoup4==4.14.2